pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aria"
version = "0.1.0"
description = "ARIA — Adaptive Reasoning & Intelligence Assistant"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "aiosqlite>=0.20.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.12",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "python-telegram-bot>=21.0",
]

[project.optional-dependencies]
voice = [
    "openai-whisper>=20231117",
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
    "pyttsx3>=2.90",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
aria = "aria.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["aria"]
